Clone `smdk test` to `fluvio sm test`
